Big commit for the Kent ACM Demo. Adding new TileSets, slick for tiledmap, guava for awesomeness. Adding new DemoModule for demo specific work. Changes to get the demo up and running, as well as improvements on position data for entities.

+++ src/com/stonetolb/engine/system/MovementSystem.java
@@ -48,8 +48,8 @@ public class MovementSystem extends EntityProcessingSystem {
 		Velocity vel = velocityMap.get(arg0);
 		Rotation rot = rotationMap.get(arg0);
 
-		pos.setX(pos.getX() + (float)((((world.getDelta() * (vel.getVelocity())) / 1000)) * Math.cos(Math.toRadians(rot.getRotation()))));
-		pos.setY(pos.getY() + (float)((((world.getDelta() * (vel.getVelocity())) / 1000)) * Math.sin(Math.toRadians(rot.getRotation()))));
+		pos.setX(pos.getX() + (int)((((world.getDelta() * (vel.getVelocity())) / 1000)) * Math.cos(Math.toRadians(rot.getRotation()))));
+		pos.setY(pos.getY() + (int)((((world.getDelta() * (vel.getVelocity())) / 1000)) * Math.sin(Math.toRadians(rot.getRotation()))));
 	}
 
 }
